opt: reorganize FK checks optbuilder code
This change moves the FK checks build code to a separate file and moves the
`addInsertionCheck` / `addDeletionCheck` code to the `fkCheckHelper`.

Release note: None
